Don't run pub/sub or spinner tests on standalone DDS (#44)
The spinner is always a ROS 2 node, and doesn't have anything to do with
the COMM value. Currently, the test is essentially a duplicate of its
associated RMW test.

For the pub/sub tests, if a standalone DDS is used for COMM, the relay
is configured to the same COMM, and ignores the `RMW_IMPLEMENTATION`
value that's being set. So the test looks like it has each RMW for a
relay, but really its none of them, and is just the standalone DDS over
and over. I really don't think we're interested in two-process
performance of standalone DDS implementations, so let's just skip them
instead of trying to fix the test to make sense.

Additionally, a static value is set for the topic used in the pub/sub
tests instead of inheriting the last value iterated in the list of
topics. This value was chosen to be large, but avoid fragmentation.

Signed-off-by: Scott K Logan <logans@cottsay.net>