Add PERF_TEST_SKIP flag to skip performance tests
This boolean flag will cause the test run for `buildfarm_perf_tests` to
skip all of the actual performance tests and only run the tests on the
package itself, such as the linters.

The idea is to use this flag to enable some continuous integration for
the package so that we can keep the linters from regressing.

Signed-off-by: Scott K Logan <logans@cottsay.net>