Unify runtime parameter
Use the same parameter, `PERF_TEST_RUNTIME`, to configure how long each
of the tests should be run.

Additionally, configure Launch to terminate any of the processes that
run for twice that duration, and if that fails, configure CTest to kill
everything after three times that duration.

Part of this change also involved setting up a proper trigger to
terminate the spinning process instead of relying on test completion to
trigger the shutdown.

Signed-off-by: Scott K Logan <logans@cottsay.net>