Create SystemMetricCollector launch action
This change moves away from the "attach" function to a pseudo-Node
launch Action that can be treated like any other launch Action.

The system_metric_collector still waits for the process under test to
start before grabbing the target PID. Unlike the previous behavior,
however, the system_metric_collector now exits as soon as the process
under test exists. Because of this, I extended the timeout value to be
several seconds longer to ensure the entire runtime of the target is
monitored.

Signed-off-by: Scott K Logan <logans@cottsay.net>