Simplify and fix log detection and cleanup
There were some regression caused by 95d1780 which resulted in an
attempt to process an empty file instead of the perf_test log. It did
not occur deterministically, and some systems produced the issue and
others did not. That is resolved by this change.

Additionally, some logs were not getting cleaned up, and the logic
around the system_metric_collector logs was simplified.

The tests now fail if results were requested but the tools wrote no
results files.

Signed-off-by: Scott K Logan <logans@cottsay.net>